refactor(renaming main and build files): renaming main files for PascalCase and moving index to/core
